refactor: StubURLSession.swift에 함께 정의되어 있던 Mock 관련 객체들 모두 분리 (bugpigg/issue-tracker#16)
