+++ iOS/IssueTracker/IssueTracker/Utility/Extension/URLSession+Extension.swift
@@ -1,0 +1,11 @@
+//
+//  URLSession+Extension.swift
+//  IssueTracker
+//
+//  Created by 김한솔 on 2022/06/20.
+//
+
+import Foundation
+
+// StubURLSession을 만들기위해, StubURLSession이 채택할 Protocol과 같은 Protocol 채택
+extension URLSession: URLSessionProtocol {}
